+++ src/reactive/effect-context.ts
@@ -1,7 +1,8 @@
 import { tag } from '../utils'
 import { asyncZone, ZoneAggregator, ZoneHistory } from '../zone'
+import { untracked, effect } from './effects'
 import { getRoot } from './registry'
-import { type EffectTrigger } from './types'
+import { cleanup, EffectAccess, ScopedCallback, stopped, type EffectTrigger } from './types'
 
 export const effectHistory = tag('effectHistory', new ZoneHistory<EffectTrigger>())
 tag('effectHistory.present', effectHistory.present)
@@ -20,4 +21,58 @@ export function isRunning(effect: EffectTrigger): boolean {
 
 export function getActiveEffect() {
 	return effectHistory.present.active
+}
+
+
+/**
+ * ADD a cleanup function to an object using the cleanup symbol.
+ * The cleanup function will be called when the object needs to be disposed.
+ * 
+ * Note: most of the time, you don't need to use this function directly.
+ * The main use if for the cleanup function to be stored with the object, as GC calls the cleanup function when the *function* is garbage collected.
+ * 
+ * @param obj - The object to attach the cleanup function to
+ * @param cleanupFn - The cleanup function to attach
+ * @returns The object with the cleanup function attached
+ */
+export function cleanedBy<T extends object>(obj: T, cleanupFn: ScopedCallback) {
+	const oldCleanup = obj[cleanup]
+	return Object.defineProperty(obj, cleanup, {
+		value: oldCleanup
+			? Object.defineProperties(
+					() => {
+						oldCleanup()
+						cleanupFn()
+					},
+					{
+						[stopped]: { get: () => oldCleanup[stopped] || cleanupFn[stopped] },
+					}
+				)
+			: cleanupFn,
+		writable: false,
+		enumerable: false,
+		configurable: true,
+	}) as T & { [cleanup]: ScopedCallback }
+}
+
+//#region greedy caching
+
+/**
+ * Creates a derived value that automatically recomputes when dependencies change
+ * @param compute - Function that computes the derived value
+ * @returns Object with value and cleanup function
+ */
+export function derived<T>(compute: (dep: EffectAccess) => T): {
+	value: T
+	[cleanup]: ScopedCallback
+} {
+	const rv = { value: undefined as unknown as T }
+	return cleanedBy(
+		rv,
+		untracked(() =>
+			effect(function derivedEffect(access) {
+				rv.value = compute(access)
+			})
+		)
+	)
 }