+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Reactive
 - `effect` now has a `named` and `opaque` "flavored" property
 - `onEffectThrow` - effect throw/catch feature
+- typings: `ScopedCallback` became `EffectTrigger` and `EffectCleanup`
 
 ## Refactoring